refactor: extract 17 pure helper modules from AgentPanel.jsx
AgentPanel.jsx had grown to 5717 lines with 121 free-standing helper
functions tangled around the 2958-line main component. None of these
helpers depended on component state, so they could be lifted into
their own files without touching the React tree.

Extract into src/component/agent/panel/ organized by concern:

  api/         _helpers, buildApiMessages, openaiMessages,
               anthropicMessages — provider-specific request shaping
  messages/    assistantMessages, toolResults, userMessage,
               imageEditTool — message-content building/serialization
  utils/       llmStats (context-usage + body-size), scheduleStatus,
               sessionPlans, streamingArgs
  input/       slashCommand, tabMention
  search/      globalSearch (cross-session search helpers)
  export/      sessionExport (markdown export pipeline)
  skills/      skillsAdapter (skill-station load + parse)

Public exports buildSessionExportMarkdown and
collectToolResultDisplayImages are re-exported from AgentPanel.jsx
so AgentPanel.export.test.jsx and AgentPanel.imageEdit.test.jsx
continue working unchanged. JSX subcomponents (PlanPanel,
ImageEditDialog, ScheduleDialog, SystemPromptDialog, CommandMenu,
StreamingArgsBubble) stay in AgentPanel.jsx for a follow-up commit.

AgentPanel.jsx: 5717 → 3966 lines (-31%). All 90 tests pass, lint clean.
