+++ src/component/agent/AgentPanel.jsx
@@ -3,9 +3,7 @@
 import { Button, Card, Dialog } from "@sunwu51/camel-ui";
 import { useEffect, useRef, useState } from "react";
 import {
-  API_TYPES,
   DEFAULT_MODEL_CONTEXT_LIMIT_TOKENS,
-  MODEL_CONTEXT_WARNING_THRESHOLD_RATIO,
   getDefaultApiType,
   normalizeApiType,
   normalizeModelContextLimitTokens,
@@ -15,7 +13,6 @@ import {
   hasDownloadsPermission,
   isMcpToolCallName
 } from "../../api/llm";
-import { connectMcpServer, listMcpResources, readMcpResource } from "../../api/mcp";
 import {
   generateSessionId,
   listSessions,
@@ -51,13 +48,12 @@ import SkillsConfig from "./SkillsConfig";
 import toast from "react-hot-toast";
 import { formatProfileForSystemPrompt } from "../../api/userProfile";
 import { refreshSessionKeywords } from "../../api/sessionKeywords";
-import { getLongToolArgumentFields } from "../../api/llm/longToolArgs";
 import {
   DEFAULT_IMAGE_MODEL,
   IMAGE_API_PROTOCOLS,
   isImageApiConfigured,
   normalizeImageApiProtocol
-} from "../../api/llm/imageApi";
+} from "../../api/llm/builtins/imageApi";
 import {
   IMAGE_REF_PATTERN,
   allocateGeneratedImageRef,
@@ -70,40 +66,98 @@ import {
 } from "./imageRefs";
 import "./chat.css";
 
+// === Extracted helper modules (see ./panel/) ===
+import {
+  isTerminalScheduleStatus,
+  formatScheduleStatus,
+  normalizeScheduleStatusClass,
+  formatRemainingSeconds,
+  getLiveRemainingSeconds
+} from "./panel/utils/scheduleStatus";
+import {
+  normalizeSessionPlans,
+  getLatestPlan,
+  normalizePlanSteps,
+  derivePlanStatus,
+  formatPlanStatus,
+  getPlanStepIcon
+} from "./panel/utils/sessionPlans";
+import { buildStreamingToolArgsState } from "./panel/utils/streamingArgs";
+import {
+  buildContextUsage,
+  getLatestContextUsageFromMessages,
+  formatModelName,
+  normalizeReasoningEffort,
+  isContextUsageWarning,
+  formatContextLimitK,
+  formatContextUsageK,
+  normalizeRequestBodySize,
+  shouldShowRequestBodySize,
+  isRequestBodySizeWarning,
+  formatRequestBodySizeM
+} from "./panel/utils/llmStats";
+import {
+  SLASH_COMMANDS,
+  shouldOpenSlashCommand,
+  filterSlashCommands,
+  buildMemoryCommandPrompt,
+  buildRecallMemoryCommandPrompt
+} from "./panel/input/slashCommand";
+import {
+  getActiveTabMentionState,
+  serializeMentionSkill,
+  filterMentionTabs,
+  queryHttpTabsForMention,
+  buildUserInjectionMeta,
+  buildInjectedUserText
+} from "./panel/input/tabMention";
+import { buildGlobalSessionSearchResult } from "./panel/search/globalSearch";
+import {
+  buildSessionExportMarkdown,
+  downloadMarkdownFile
+} from "./panel/export/sessionExport";
+import {
+  buildFinalAssistantMessage,
+  buildAssistantToolCallMessage
+} from "./panel/messages/assistantMessages";
+import {
+  buildToolResultMessages,
+  buildLlmErrorDisplayMessage,
+  stampLastUserDuration,
+  buildDisplayToolResultMessage,
+  collectToolResultDisplayImages
+} from "./panel/messages/toolResults";
+import {
+  findImageEditMcpTool,
+  getMcpToolCallName,
+  buildImageEditUserPrompt,
+  buildImageEditMessageRefs
+} from "./panel/messages/imageEditTool";
+import {
+  imageBlockToDataUrl,
+  isImageFile,
+  getClipboardImageFiles,
+  imageFileToAttachmentItem,
+  resolveImageRefsInValue,
+  normalizeImageRefToken,
+  summarizeImageRefCache,
+  buildUserMessageContent
+} from "./panel/messages/userMessage";
+import { buildApiMessages, buildPlatformSystemPrompt } from "./panel/api/buildApiMessages";
+import {
+  mergeMcpToolLists,
+  loadSkillsIndexFromSkillStation,
+  loadSkillStationTools
+} from "./panel/skills/skillsAdapter";
+
+// Re-export for tests (AgentPanel.export.test.jsx imports these from this file)
+export { buildSessionExportMarkdown, collectToolResultDisplayImages };
+
 const SYSTEM_PROMPT_PLACEHOLDER =
   "例如：你是一位情感大师，擅长共情、倾听和温柔地拆解亲密关系问题。回答时先复述用户感受，再给出具体可执行的沟通建议；避免评判，语气温暖、真诚、稳定。";
 const CHAT_AUTO_FOLLOW_BOTTOM_THRESHOLD_PX = 80;
 const SESSION_KEYWORDS_REFRESH_INTERVAL_MS = 3 * 60 * 1000;
-const IMAGE_DEREF_PATTERN = /^\|deRef:(img_[A-Za-z0-9_-]+)\|$/;
 const IMAGE_REFS_DEBUG_GLOBAL = "__tabManagerImageRefs";
-const IMAGE_FILE_EXT = /\.(png|jpe?g|gif|webp|bmp)$/i;
-const SLASH_COMMANDS = [
-  {
-    id: "mem",
-    name: "/mem",
-    title: "总结到记忆",
-    description: "提炼本次对话中对未来有用的信息，并在有记忆工具时保存。"
-  },
-  {
-    id: "recall_mem",
-    name: "/recall_mem",
-    title: "召回相关记忆",
-    description: "根据当前对话检索长期记忆，并把相关信息拉取到当前上下文。"
-  },
-  {
-    id: "clear",
-    name: "/clear",
-    title: "清空当前会话",
-    description: "与工具栏清空按钮相同，会清空消息、计划和关键词。"
-  }
-];
-
-/**
- * Main Agent chat panel with session management.
- * - Auto-saves conversation to chrome.storage.local
- * - Toolbar at top: new session / title / history dropdown
- * - Restores last session on mount
- */
 export default function AgentPanel() {
   const [messages, setMessages] = useState([]);
   const [input, setInput] = useState("");
@@ -3066,9 +3120,6 @@ export default function AgentPanel() {
 // ==================== Helper functions ====================
 
 
-function normalizeSessionPlans(plans) {
-  return Array.isArray(plans) ? plans.filter(Boolean) : [];
-}
 
 function useAutoScrollMenuOnPointerEdge(active) {
   const ref = useRef(null);
@@ -3217,261 +3268,10 @@ function InputCommandMenu({
 }
 /* eslint-enable react/prop-types */
 
-function shouldOpenSlashCommand(input) {
-  return /^\/[a-zA-Z0-9_-]*$/.test(String(input || "").trimStart());
-}
 
-function filterSlashCommands(commands, skills, selectedSkills, input) {
-  const query = String(input || "").trimStart().replace(/^\//, "").toLowerCase();
-  const selectedSkillPaths = new Set((selectedSkills || []).map(skill => skill.path));
-  const builtins = (commands || [])
-    .filter(command => {
-      if (!query) return true;
-      return command.id.includes(query) || command.name.toLowerCase().includes(query) || command.title.toLowerCase().includes(query);
-    });
-  const skillCommands = (skills || [])
-    .map(serializeMentionSkill)
-    .filter(skill => skill.path && !selectedSkillPaths.has(skill.path))
-    .filter(skill => {
-      if (!query) return true;
-      return `${skill.name || ""} ${skill.path || ""} ${skill.description || ""}`.toLowerCase().includes(query);
-    })
-    .slice(0, 20)
-    .map(skill => ({
-      id: `skill:${skill.path}`,
-      type: "skill",
-      skill,
-      name: `/${skill.name || skill.path}`,
-      title: skill.name || skill.path,
-      description: skill.description || `使用 skill: ${skill.path}`
-    }));
-  return [...builtins, ...skillCommands];
-}
 
-function getActiveTabMentionState(input) {
-  const text = String(input || "");
-  const match = text.match(/(^|\s)@([^\s@]*)$/);
-  if (!match) return null;
-  const atOffset = match[0].lastIndexOf("@");
-  const start = match.index + atOffset;
-  return {
-    start,
-    end: text.length,
-    query: match[2] || ""
-  };
-}
 
-function isHttpTab(tab) {
-  return typeof tab?.id === "number" && /^https?:\/\//i.test(tab?.url || "");
-}
 
-function serializeMentionTab(tab) {
-  return {
-    id: tab.id,
-    url: tab.url || "",
-    title: tab.title || "",
-    windowId: tab.windowId,
-    favIconUrl: tab.favIconUrl || ""
-  };
-}
-
-function serializeMentionSkill(skill) {
-  return {
-    path: String(skill?.path || "").trim(),
-    name: String(skill?.name || "").trim(),
-    description: String(skill?.description || "").trim()
-  };
-}
-
-function filterMentionTabs(tabs, query, selectedTabs) {
-  const selectedIds = new Set((selectedTabs || []).map(tab => tab.id));
-  const normalizedQuery = String(query || "").trim().toLowerCase();
-  return (tabs || [])
-    .filter(tab => !selectedIds.has(tab.id))
-    .filter(tab => {
-      if (!normalizedQuery) return true;
-      return `${tab.title || ""} ${tab.url || ""} ${tab.id}`.toLowerCase().includes(normalizedQuery);
-    })
-    .slice(0, 12);
-}
-
-async function queryHttpTabsForMention() {
-  if (!chrome?.tabs?.query) return [];
-  const tabs = await chrome.tabs.query({});
-  return (Array.isArray(tabs) ? tabs : []).filter(isHttpTab).map(serializeMentionTab);
-}
-
-function buildUserInjectionMeta(tabs, skills) {
-  const normalizedTabs = (tabs || []).map(serializeMentionTab).filter(tab => tab.id);
-  const normalizedSkills = (skills || []).map(serializeMentionSkill).filter(skill => skill.path);
-  if (normalizedTabs.length === 0 && normalizedSkills.length === 0) return null;
-  return {
-    tabs: normalizedTabs,
-    skills: normalizedSkills
-  };
-}
-
-function buildInjectedUserText(text, injectionMeta) {
-  const body = String(text || "").trim() || "请根据我指定的上下文回答。";
-  const tabLines = (injectionMeta?.tabs || []).map(tab =>
-    `- tabId: ${tab.id}; title: ${tab.title || "未命名标签页"}; url: ${tab.url}`
-  );
-  const skillLines = (injectionMeta?.skills || []).map(skill =>
-    `- directoryName: ${skill.path}; name: ${skill.name || skill.path}; description: ${skill.description || ""}`
-  );
-  const parts = [body];
-  if (tabLines.length > 0) {
-    parts.push(
-      "",
-      "用户在输入框中明确 @ 选择了以下已打开的浏览器标签页。回答时应优先把这些 tabId 视为用户指定目标；如需读取页面内容，请直接对对应 tabId 使用 tab_extract 等浏览器工具：",
-      ...tabLines
-    );
-  }
-  if (skillLines.length > 0) {
-    parts.push(
-      "",
-      "用户在输入框中明确选择了以下 skill。请优先使用这些 skill 来回答接下来的问题或完成任务；需要完整 workflow 时，使用 skill-bridge 的 get_skill_detail 读取对应 directoryName：",
-      ...skillLines
-    );
-  }
-  return parts.join("\n");
-}
-
-function buildMemoryCommandPrompt() {
-  return [
-    "请总结本次对话中对未来有用、稳定、值得长期保存的信息。",
-    "如果当前可用工具中存在长期记忆/保存记忆能力，请使用相应工具保存这些信息；工具名称可能被 MCP 或系统加前缀，请根据工具描述判断。",
-    "只保存用户偏好、长期项目背景、稳定决策、反复会用到的上下文或纠错后的规则。不要保存临时状态、一次性问题、明显过期的信息或敏感凭据。",
-    "完成后请简要告诉我保存了什么；如果没有可用的记忆保存工具，或者没有值得保存的信息，也请明确说明。"
-  ].join("\n");
-}
-
-function buildRecallMemoryCommandPrompt() {
-  return [
-    "请根据当前对话、当前任务、用户最近的问题和已知项目上下文，检索可能相关的长期记忆。",
-    "如果当前可用工具中存在长期记忆检索/召回/搜索能力，请使用相应工具拉取相关记忆；工具名称可能被 MCP 或系统加前缀，请根据工具名称和描述判断。",
-    "检索查询应简洁，包含当前任务里的关键项目名、模块、功能、用户偏好、决策或约束。不要为临时网页状态、一次性事实或最新信息使用记忆召回。",
-    "完成后请把召回到的相关记忆简要整理到当前上下文，并说明哪些内容会影响接下来的回答或执行；如果没有可用的记忆检索工具或没有相关记忆，也请明确说明。"
-  ].join("\n");
-}
-
-function buildGlobalSessionSearchResult(sessionEntry, messages, query) {
-  const normalizedQuery = String(query || "").trim().toLowerCase();
-  if (!sessionEntry || !normalizedQuery) return null;
-  let hitCount = 0;
-  let firstSnippet = "";
-  let lastMessageTime = null;
-  for (const message of messages || []) {
-    if (!isGlobalSearchableMessage(message)) continue;
-    const timestamp = message.sentAt || message.completedAt || message.updatedAt || null;
-    if (timestamp && (!lastMessageTime || timestamp > lastMessageTime)) lastMessageTime = timestamp;
-    const text = getGlobalSearchableMessageText(message);
-    if (!text) continue;
-    const lowerText = text.toLowerCase();
-    let fromIndex = 0;
-    while (fromIndex < lowerText.length) {
-      const foundAt = lowerText.indexOf(normalizedQuery, fromIndex);
-      if (foundAt < 0) break;
-      hitCount += 1;
-      if (!firstSnippet) firstSnippet = buildSearchSnippet(text, foundAt, normalizedQuery.length);
-      fromIndex = foundAt + Math.max(1, normalizedQuery.length);
-    }
-  }
-  if (hitCount === 0) return null;
-  return {
-    sessionId: sessionEntry.id,
-    title: sessionEntry.title || "新会话",
-    startedAt: sessionEntry.startedAt || 0,
-    updatedAt: sessionEntry.updatedAt || 0,
-    lastMessageTime,
-    hitCount,
-    snippet: firstSnippet || "命中当前关键词"
-  };
-}
-
-function buildStreamingToolArgsState(event) {
-  const name = event?.name || "";
-  const rawArgs = typeof event?.arguments === "string" ? event.arguments : "";
-  const fields = getLongToolArgumentFields(name);
-  if (!name || fields.length === 0) return null;
-  const preview = buildStreamingToolArgumentPreview(name, rawArgs, fields);
-  return {
-    id: event?.id || event?.responseItemId || `${name}-${event?.index ?? 0}`,
-    name,
-    preview
-  };
-}
-
-function buildStreamingToolArgumentPreview(toolName, rawArgs, fields) {
-  const parsed = tryParseJson(rawArgs);
-  if (parsed && typeof parsed === "object") {
-    const parts = [];
-    for (const field of fields) {
-      if (typeof parsed[field] === "string" && parsed[field]) {
-        parts.push(`${field}=${parsed[field]}`);
-      }
-    }
-    if (parts.length > 0) return parts.join("\n");
-  }
-
-  const extracted = [];
-  for (const field of fields) {
-    const value = extractPartialJsonStringValue(rawArgs, field);
-    if (value) extracted.push(`${field}=${value}`);
-  }
-  if (extracted.length > 0) return extracted.join("\n");
-  return rawArgs || `${toolName} 参数生成中`;
-}
-
-function tryParseJson(text) {
-  try {
-    return JSON.parse(text || "{}");
-  } catch {
-    return null;
-  }
-}
-
-function extractPartialJsonStringValue(jsonText, fieldName) {
-  const text = String(jsonText || "");
-  const fieldPattern = new RegExp(`"${escapeRegExp(fieldName)}"\\s*:\\s*"`, "g");
-  const match = fieldPattern.exec(text);
-  if (!match) return "";
-  let result = "";
-  let escaped = false;
-  for (let i = match.index + match[0].length; i < text.length; i++) {
-    const ch = text[i];
-    if (escaped) {
-      result += decodeJsonEscapeChar(ch);
-      escaped = false;
-      continue;
-    }
-    if (ch === "\\") {
-      escaped = true;
-      continue;
-    }
-    if (ch === "\"") break;
-    result += ch;
-  }
-  return result;
-}
-
-function decodeJsonEscapeChar(ch) {
-  switch (ch) {
-    case "n": return "\n";
-    case "r": return "\r";
-    case "t": return "\t";
-    case "b": return "\b";
-    case "f": return "\f";
-    case "\"": return "\"";
-    case "\\": return "\\";
-    case "/": return "/";
-    default: return ch;
-  }
-}
-
-function escapeRegExp(text) {
-  return String(text).replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
 
 /* eslint-disable react/prop-types */
 function StreamingToolArgsBubble({ state }) {
@@ -3485,93 +3285,6 @@ function StreamingToolArgsBubble({ state }) {
   );
 }
 
-function isGlobalSearchableMessage(message) {
-  if (!message) return false;
-  if (message.role !== "user" && message.role !== "assistant") return false;
-  if (message.role === "assistant" && Array.isArray(message.tool_calls) && message.tool_calls.length > 0) return false;
-  if (Array.isArray(message.content) && message.content.some(block => block?.type === "tool_use" || block?.type === "tool_result")) {
-    return false;
-  }
-  return true;
-}
-
-function getGlobalSearchableMessageText(message) {
-  const { content } = message || {};
-  if (typeof content === "string") return content;
-  if (!Array.isArray(content)) return "";
-  return content
-    .map(block => {
-      if (typeof block === "string") return block;
-      if (block?.type === "text" && typeof block.text === "string") return block.text;
-      return "";
-    })
-    .filter(Boolean)
-    .join("\n");
-}
-
-function buildSearchSnippet(text, hitStart, queryLength) {
-  const source = String(text || "").replace(/\s+/g, " ").trim();
-  if (!source) return "";
-  const safeStart = Math.max(0, Math.min(source.length, hitStart));
-  const start = Math.max(0, safeStart - 36);
-  const end = Math.min(source.length, safeStart + queryLength + 56);
-  return `${start > 0 ? "…" : ""}${source.slice(start, end)}${end < source.length ? "…" : ""}`;
-}
-
-function getLatestPlan(plans) {
-  const normalized = normalizeSessionPlans(plans);
-  return normalized.length > 0 ? normalized[normalized.length - 1] : null;
-}
-
-function normalizePlanSteps(steps) {
-  if (!Array.isArray(steps)) return [];
-  return steps
-    .map((step, index) => ({
-      id: step?.id || `step_${index + 1}`,
-      title: String(step?.title || "").trim(),
-      description: String(step?.description || "").trim(),
-      status: normalizePlanStepStatus(step?.status || "pending"),
-      note: String(step?.note || "").trim(),
-      updatedAt: step?.updatedAt || null
-    }))
-    .filter(step => step.title);
-}
-
-function normalizePlanStepStatus(status) {
-  const value = String(status || "pending");
-  return ["pending", "in_progress", "completed", "blocked", "skipped"].includes(value) ? value : "pending";
-}
-
-function derivePlanStatus(steps) {
-  const list = Array.isArray(steps) ? steps : [];
-  if (list.length === 0) return "draft";
-  if (list.some(step => step.status === "blocked")) return "blocked";
-  if (list.every(step => step.status === "completed" || step.status === "skipped")) return "completed";
-  if (list.some(step => step.status === "in_progress" || step.status === "completed" || step.status === "skipped")) return "in_progress";
-  return "approved";
-}
-
-function formatPlanStatus(status) {
-  switch (status) {
-    case "draft": return "待确认";
-    case "approved": return "已确认";
-    case "in_progress": return "执行中";
-    case "completed": return "已完成";
-    case "blocked": return "受阻";
-    case "cancelled": return "已取消";
-    default: return status || "计划";
-  }
-}
-
-function getPlanStepIcon(status) {
-  switch (status) {
-    case "completed": return "✅";
-    case "in_progress": return "🔄";
-    case "blocked": return "⛔";
-    case "skipped": return "⏭️";
-    default: return "⬜";
-  }
-}
 
 /* eslint-disable react/prop-types */
 function SessionPlanPanel({ plan, collapsed = false, onToggleCollapsed }) {
@@ -4014,144 +3727,6 @@ export function ImageEditDialog({ request, disabled = false, onCancel, onConfirm
   );
 }
 
-function buildContextUsage(apiType, model, usage) {
-  if (!usage || typeof usage !== "object") return null;
-  const tokens = calculateContextTokens(apiType, usage);
-  return {
-    apiType: normalizeApiType(apiType || getDefaultApiType()),
-    model: model || "",
-    tokens: Number.isFinite(tokens) ? tokens : null,
-    usageStatus: Number.isFinite(tokens) ? "ok" : "unrecognized",
-    usage
-  };
-}
-
-function getLatestContextUsageFromMessages(messages, fallbackConfig = {}) {
-  for (let index = (messages || []).length - 1; index >= 0; index--) {
-    const msg = messages[index];
-    if (!msg?.usage) continue;
-    const usageInfo = buildContextUsage(
-      normalizeApiType(msg._usageApiType || fallbackConfig.apiType || getDefaultApiType()),
-      msg._usageModel || fallbackConfig.model || "",
-      msg.usage
-    );
-    if (usageInfo) return usageInfo;
-  }
-  return null;
-}
-
-function calculateContextTokens(apiType, usage) {
-  if (!usage || typeof usage !== "object") return null;
-  const anthropicFields = [
-    "input_tokens",
-    "output_tokens",
-    "cache_read_input_tokens",
-    "cache_creation_input_tokens"
-  ];
-  const openAiFields = ["prompt_tokens", "completion_tokens"];
-
-  if (normalizeApiType(apiType) === API_TYPES.ANTHROPIC) {
-    return firstFiniteNumber(
-      sumTokenFields(usage, anthropicFields),
-      sumTokenFields(usage, openAiFields),
-      getFirstUsageNumber(usage, ["total_tokens", "totalTokens", "total"])
-    );
-  }
-  return firstFiniteNumber(
-    sumTokenFields(usage, openAiFields),
-    sumTokenFields(usage, anthropicFields),
-    getFirstUsageNumber(usage, ["total_tokens", "totalTokens", "total"])
-  );
-}
-
-function sumTokenFields(source, fields) {
-  let total = 0;
-  let hasValue = false;
-  for (const field of fields) {
-    const value = Number(source?.[field]);
-    if (!Number.isFinite(value)) continue;
-    total += value;
-    hasValue = true;
-  }
-  return hasValue ? total : null;
-}
-
-function firstFiniteNumber(...values) {
-  for (const value of values) {
-    if (Number.isFinite(value)) return value;
-  }
-  return null;
-}
-
-function getFirstUsageNumber(source, fields) {
-  for (const field of fields) {
-    const value = Number(source?.[field]);
-    if (Number.isFinite(value)) return value;
-  }
-  return null;
-}
-
-function formatModelName(model) {
-  return String(model || "").trim() || "未配置";
-}
-
-function normalizeReasoningEffort(value) {
-  return ["default", "low", "medium", "high", "xhigh"].includes(value) ? value : "default";
-}
-
-function isContextUsageWarning(contextUsage, limitTokens) {
-  const tokens = Number(contextUsage?.tokens);
-  const normalizedLimit = normalizeModelContextLimitTokens(limitTokens);
-  return Number.isFinite(tokens) && tokens >= normalizedLimit * MODEL_CONTEXT_WARNING_THRESHOLD_RATIO;
-}
-
-function formatContextLimitK(limitTokens) {
-  const normalizedLimit = normalizeModelContextLimitTokens(limitTokens);
-  if (normalizedLimit >= 1000000) return `${normalizedLimit / 1000000}M`;
-  return `${Math.round(normalizedLimit / 1000)}K`;
-}
-
-function formatContextUsageK(contextUsage) {
-  const tokens = Number(contextUsage?.tokens);
-  if (contextUsage?.usageStatus === "unrecognized") return "未识别";
-  if (!Number.isFinite(tokens)) return "未返回";
-  const value = tokens / 1000;
-  if (value >= 100) return `${Math.round(value)}K`;
-  if (value >= 10) return `${value.toFixed(1)}K`;
-  return `${value.toFixed(2)}K`;
-}
-
-const REQUEST_BODY_SIZE_DISPLAY_BYTES = 1024 * 1024;
-const REQUEST_BODY_SIZE_WARNING_BYTES = 5 * 1024 * 1024;
-
-function normalizeRequestBodySize(size) {
-  const bytes = Number(size?.bytes ?? size);
-  if (!Number.isFinite(bytes) || bytes < 0) return null;
-  return {
-    bytes,
-    apiType: size?.apiType || "",
-    model: size?.model || ""
-  };
-}
-
-function shouldShowRequestBodySize(size) {
-  const bytes = Number(size?.bytes);
-  return Number.isFinite(bytes) && bytes >= REQUEST_BODY_SIZE_DISPLAY_BYTES;
-}
-
-function isRequestBodySizeWarning(size) {
-  const bytes = Number(size?.bytes);
-  return Number.isFinite(bytes) && bytes >= REQUEST_BODY_SIZE_WARNING_BYTES;
-}
-
-function formatRequestBodySizeM(size) {
-  const bytes = Number(size?.bytes);
-  if (!Number.isFinite(bytes)) return "未计算";
-  const value = bytes / REQUEST_BODY_SIZE_DISPLAY_BYTES;
-  if (value >= 10) return `${Math.round(value)}M`;
-  return `${value.toFixed(1)}M`;
-}
-
 /* eslint-disable react/prop-types */
 function SessionSystemPromptDialogBody({ initialValue = "", initiallyApplyToNewSessions = false, onSave }) {
   const [draft, setDraft] = useState(initialValue || "");
@@ -4375,1343 +3950,17 @@ function ScheduleJobsDialogBody() {
   );
 }
 
-function isTerminalScheduleStatus(status) {
-  return status === "succeeded" || status === "failed" || status === "cancelled";
-}
 
-export function buildSessionExportMarkdown({ title, sessionId, messages }) {
-  const imageRefMap = buildExportImageRefMap(messages);
-  const sections = [
-    `# ${title || "新会话"}`,
-    "",
-    `- 导出时间: ${new Date().toLocaleString()}`,
-    `- 会话 ID: ${sessionId || ""}`,
-    ""
-  ];
 
-  for (const msg of messages || []) {
-    sections.push(...serializeExportMessage(msg, imageRefMap));
-  }
 
-  return sections.join("\n").replace(/\n{3,}/g, "\n\n").trim();
-}
 
-function serializeExportMessage(msg, imageRefMap = new Map()) {
-  if (!msg || !msg.role) return [];
 
-  if (msg.role === "user") {
-    return serializeUserExportMessage(msg);
-  }
 
-  if (msg.role === "assistant") {
-    return serializeAssistantExportMessage(msg, imageRefMap);
-  }
 
-  if (msg.role === "tool") {
-    return [
-      `## 工具结果${msg.tool_name ? ` · ${msg.tool_name}` : ""}`,
-      "",
-      formatToolResultForMarkdown(msg),
-      ""
-    ];
-  }
 
-  if (msg.role === "error") {
-    return [
-      "## 错误",
-      "",
-      formatJsonFence(msg.content ?? {}),
-      ""
-    ];
-  }
 
-  return [
-    `## ${msg.role}`,
-    "",
-    formatUnknownContentForMarkdown(msg.content),
-    ""
-  ];
-}
 
-function serializeUserExportMessage(msg) {
-  return [
-    "---",
-    "",
-    "## 用户",
-    "",
-    formatUserContentForMarkdown(msg.content),
-    ""
-  ];
-}
 
-function formatUserContentForMarkdown(content) {
-  if (typeof content === "string") return content.trim() || "_空内容_";
-  if (!Array.isArray(content)) return formatUnknownContentForMarkdown(content);
 
-  const parts = content
-    .map(formatUserContentBlockForMarkdown)
-    .filter(part => typeof part === "string" && part.trim().length > 0);
 
-  return parts.length > 0 ? parts.join("\n\n") : "_空内容_";
-}
 
-function formatUserContentBlockForMarkdown(block) {
-  if (!block || typeof block !== "object") return "";
-
-  if (block.type === "text") {
-    return String(block.text ?? "").trim();
-  }
-
-  if (block.type === "file") {
-    const fileName = String(block.fileName || block.name || "attachment.txt").trim();
-    return [`### 附件 · ${fileName}`, "", formatTextFence(block.text ?? "")].join("\n");
-  }
-
-  const dataUrl = imageBlockToDataUrl(block);
-  if (dataUrl) {
-    const mediaType = parseImageDataUrl(dataUrl)?.mediaType || "image";
-    return `![用户图片 · ${mediaType}](${dataUrl})`;
-  }
-
-  return formatJsonFence(block);
-}
-
-function serializeAssistantExportMessage(msg, imageRefMap = new Map()) {
-  const sections = [];
-
-  if (typeof msg.content === "string" && msg.content.trim()) {
-    sections.push("## 助手", "", replaceMarkdownImageDerefSources(msg.content.trim(), imageRefMap), "");
-  }
-
-  if (Array.isArray(msg.content)) {
-    for (const block of msg.content) {
-      if (!block) continue;
-      if (block.type === "text" && block.text) {
-        sections.push("## 助手", "", replaceMarkdownImageDerefSources(String(block.text).trim(), imageRefMap), "");
-      } else if (block.type === "tool_use") {
-        sections.push(
-          `## 工具调用${block.name ? ` · ${block.name}` : ""}`,
-          "",
-          formatJsonFence(block.input ?? {}),
-          ""
-        );
-      }
-    }
-  }
-
-  if (Array.isArray(msg.tool_calls)) {
-    for (const toolCall of msg.tool_calls) {
-      const toolName = toolCall?.function?.name || toolCall?.name || "tool";
-      let toolArgs = toolCall?.function?.arguments ?? toolCall?.arguments ?? toolCall?.args ?? {};
-      if (typeof toolArgs === "string") {
-        try {
-          toolArgs = JSON.parse(toolArgs);
-        } catch (error) {
-          toolArgs = { raw: toolArgs };
-        }
-      }
-      sections.push(
-        `## 工具调用 · ${toolName}`,
-        "",
-        formatJsonFence(toolArgs),
-        ""
-      );
-    }
-  }
-
-  if (sections.length === 0) {
-    sections.push("## 助手", "", "_空内容_", "");
-  }
-
-  return sections;
-}
-
-function buildExportImageRefMap(messages = []) {
-  const refs = new Map();
-  for (const msg of messages || []) {
-    for (const item of normalizeMessageImageRefs(msg?.imageRefs)) {
-      if (isBase64DataUrl(item.dataUrl) && !refs.has(item.ref)) {
-        refs.set(item.ref, item.dataUrl);
-      }
-    }
-  }
-  return refs;
-}
-
-function replaceMarkdownImageDerefSources(markdown, imageRefMap = new Map()) {
-  if (typeof markdown !== "string" || !markdown.includes("|deRef:")) return markdown;
-  return markdown.replace(/(!\[[^\]\n]*\]\()\|deRef:(img_[A-Za-z0-9_-]+)\|(\))/g, (match, prefix, ref, suffix) => {
-    const dataUrl = imageRefMap.get(ref);
-    return isBase64DataUrl(dataUrl) ? `${prefix}${dataUrl}${suffix}` : match;
-  });
-}
-
-function formatToolResultForMarkdown(msg) {
-  const parsed = parseToolMessageContent(msg.content);
-  const contentBlock = typeof parsed === "string"
-    ? formatTextFence(parsed)
-    : formatJsonFence(parsed ?? {});
-  const displayImages = Array.isArray(msg.displayImages) && msg.displayImages.length > 0
-    ? msg.displayImages.map(image => image?.url).filter(Boolean)
-    : (msg.displayImageUrl ? [msg.displayImageUrl] : []);
-
-  if (displayImages.length === 0) {
-    return contentBlock;
-  }
-
-  return [
-    contentBlock,
-    "",
-    ...displayImages.map((url, index) => `![工具图片${displayImages.length > 1 ? ` ${index + 1}` : ""}](${url})`)
-  ].join("\n");
-}
-
-function formatUnknownContentForMarkdown(content) {
-  if (typeof content === "string") return content.trim() || "_空内容_";
-  if (Array.isArray(content)) return formatJsonFence(content);
-  if (content && typeof content === "object") return formatJsonFence(content);
-  return "_空内容_";
-}
-
-function formatScheduleStatus(status) {
-  switch (status) {
-    case "pending": return "待执行";
-    case "running": return "执行中";
-    case "succeeded": return "已成功";
-    case "failed": return "已失败";
-    case "cancelled": return "已取消";
-    default: return status || "未知";
-  }
-}
-
-function normalizeScheduleStatusClass(status) {
-  switch (status) {
-    case "pending":
-    case "running":
-    case "succeeded":
-    case "failed":
-    case "cancelled":
-      return status;
-    default:
-      return "unknown";
-  }
-}
-
-function formatRemainingSeconds(seconds) {
-  const total = Math.max(0, Number(seconds) || 0);
-  const hours = Math.floor(total / 3600);
-  const minutes = Math.floor((total % 3600) / 60);
-  const secs = total % 60;
-  if (hours > 0) return `${hours}小时 ${minutes}分`;
-  if (minutes > 0) return `${minutes}分 ${secs}秒`;
-  return `${secs}秒`;
-}
-
-function getLiveRemainingSeconds(job, now = Date.now()) {
-  if (!job || job.status !== "pending") return 0;
-  const fireAtMs = job.fireAt ? new Date(job.fireAt).getTime() : NaN;
-  if (Number.isFinite(fireAtMs)) {
-    return Math.max(0, Math.round((fireAtMs - now) / 1000));
-  }
-  return Math.max(0, Number(job.remainingSeconds) || 0);
-}
-
-function formatJsonFence(value) {
-  let text = "";
-  try {
-    text = JSON.stringify(value, null, 2);
-  } catch (error) {
-    text = String(value ?? "");
-  }
-  return `\`\`\`json\n${text}\n\`\`\``;
-}
-
-function formatTextFence(value) {
-  return `\`\`\`text\n${String(value ?? "")}\n\`\`\``;
-}
-
-function buildFinalAssistantMessage(apiType, model, textContent, doneMsg = {}) {
-  if (apiType === "anthropic" && Array.isArray(doneMsg.content)) {
-    return copyAssistantUsageFields(apiType, model, doneMsg, copyAnthropicThinkingFields(doneMsg, { role: "assistant", content: doneMsg.content }));
-  }
-
-  const message = {
-    role: "assistant",
-    content: textContent || doneMsg.content || ""
-  };
-  if (doneMsg?.response_id) {
-    message._responsesResponseId = doneMsg.response_id;
-  }
-  if (Array.isArray(doneMsg?.response_content) && doneMsg.response_content.length > 0) {
-    message._responsesContent = doneMsg.response_content;
-  }
-  if (Array.isArray(doneMsg?.response_reasoning_items) && doneMsg.response_reasoning_items.length > 0) {
-    message._responsesReasoningItems = doneMsg.response_reasoning_items;
-  }
-  return copyAssistantUsageFields(apiType, model, doneMsg, copyAssistantReasoningFields(doneMsg, message));
-}
-
-function downloadMarkdownFile(filename, markdown) {
-  const safeFilename = String(filename || "session.md").trim() || "session.md";
-  const blob = new Blob([String(markdown ?? "")], { type: "text/markdown;charset=utf-8" });
-  const url = URL.createObjectURL(blob);
-  const anchor = document.createElement("a");
-  anchor.href = url;
-  anchor.download = safeFilename;
-  anchor.style.display = "none";
-  document.body.appendChild(anchor);
-  try {
-    anchor.click();
-    return { success: true, fileName: safeFilename, size: blob.size, source: "side-panel-blob" };
-  } finally {
-    anchor.remove();
-    // Keep the blob URL alive long enough for Chromium to start consuming it.
-    window.setTimeout(() => URL.revokeObjectURL(url), 1000);
-  }
-}
-
-function buildAssistantToolCallMessage(apiType, model, textContent, doneMsg) {
-  if (normalizeApiType(apiType) === API_TYPES.ANTHROPIC) {
-    return copyAssistantUsageFields(apiType, model, doneMsg, copyAnthropicThinkingFields(doneMsg, { role: "assistant", content: doneMsg.content }));
-  }
-
-  const message = {
-    role: "assistant",
-    content: textContent || null,
-    tool_calls: doneMsg._openaiToolCalls
-  };
-  if (doneMsg?.response_id) {
-    message._responsesResponseId = doneMsg.response_id;
-  }
-  if (Array.isArray(doneMsg?.response_content) && doneMsg.response_content.length > 0) {
-    message._responsesContent = doneMsg.response_content;
-  }
-  if (Array.isArray(doneMsg?.response_reasoning_items) && doneMsg.response_reasoning_items.length > 0) {
-    message._responsesReasoningItems = doneMsg.response_reasoning_items;
-  }
-  return copyAssistantUsageFields(apiType, model, doneMsg, copyAssistantReasoningFields(doneMsg, message));
-}
-
-function copyAssistantUsageFields(apiType, model, source, target) {
-  if (source?.usage && typeof source.usage === "object") {
-    target.usage = source.usage;
-    target._usageApiType = apiType || "";
-    target._usageModel = model || "";
-  }
-  return target;
-}
-
-function copyAnthropicThinkingFields(source, target) {
-  for (const field of ["thinking_blocks", "provider_specific_fields"]) {
-    const value = source?.[field];
-    if (value == null) continue;
-    if (Array.isArray(value) && value.length === 0) continue;
-    target[field] = value;
-  }
-  return target;
-}
-
-function copyAssistantReasoningFields(source, target) {
-  for (const field of ["reasoning_content", "reasoning", "reasoning_details", "thinking"]) {
-    const value = source?.[field];
-    if (value == null) continue;
-    if (typeof value === "string" && value.length === 0) continue;
-    if (Array.isArray(value) && value.length === 0) continue;
-    target[field] = value;
-  }
-  return target;
-}
-
-function buildToolResultMessages(toolResults, targetSessionId, registerImageDataUrl) {
-  return toolResults.map(tr => buildDisplayToolResultMessage(tr, targetSessionId, registerImageDataUrl));
-}
-
-function buildLlmErrorDisplayMessage(error) {
-  const code = error?.code || "LLM_ERROR";
-  const message = error?.message || "LLM 请求失败";
-  const failures = Array.isArray(error?.failures) ? error.failures : [];
-  return {
-    role: "error",
-    content: {
-      code,
-      message,
-      status: error?.status || null,
-      attempts: Number(error?.attempts) || failures.length || 1,
-      maxAttempts: Number(error?.maxAttempts) || failures.length || 1,
-      apiType: error?.apiType || "",
-      failures,
-      detail: error?.detail || null
-    }
-  };
-}
-
-function stampLastUserDuration(messages) {
-  const now = Date.now();
-  const updated = [...messages];
-  for (let i = updated.length - 1; i >= 0; i--) {
-    if (updated[i].role === "user" && typeof updated[i].sentAt === "number") {
-      updated[i] = { ...updated[i], durationMs: now - updated[i].sentAt };
-      break;
-    }
-  }
-  return updated;
-}
-
-function buildDisplayToolResultMessage(toolResult, targetSessionId, registerImageDataUrl) {
-  const displayImages = collectToolResultDisplayImages(toolResult?.result);
-  const primaryImage = displayImages[0] || null;
-  const parsedImage = parseImageDataUrl(primaryImage?.url);
-  const imageRefs = [];
-  if (targetSessionId && typeof registerImageDataUrl === "function") {
-    for (const image of displayImages) {
-      if (!isBase64DataUrl(image.url)) continue;
-      const ref = registerImageDataUrl(targetSessionId, image.url);
-      if (ref) imageRefs.push({ ref, dataUrl: image.url });
-    }
-  }
-  const summary = summarizeToolResult(toolResult.result, imageRefs);
-  const serializedContent = serializeToolResult(summary);
-  const message = {
-    role: "tool",
-    tool_call_id: toolResult.id,
-    tool_name: toolResult.name,
-    content: serializedContent,
-    displayImageUrl: primaryImage?.url || undefined,
-    displayImages: displayImages.length > 0 ? displayImages : undefined,
-    displayImageMediaType: parsedImage?.mediaType,
-    durationMs: typeof toolResult.durationMs === "number" ? toolResult.durationMs : undefined,
-  };
-  if (imageRefs.length > 0) {
-    message.imageRefs = imageRefs;
-  }
-  return message;
-}
-
-function serializeToolResult(summary) {
-  const json = JSON.stringify(summary);
-  if (typeof json === "string") return json;
-  return JSON.stringify(normalizeToolSummary(summary));
-}
-
-function summarizeToolResult(result, imageRefs = []) {
-  if (!result || typeof result !== "object") return result;
-
-  const summary = { ...result };
-  if (typeof summary.dataUrl === "string" && summary.dataUrl.startsWith("data:")) {
-    delete summary.dataUrl;
-    summary.imageOmittedFromTextContext = true;
-  }
-  if (imageRefs.length > 0) {
-    summary.imageRefs = imageRefs.map(({ ref }) => ref);
-    summary.imageRefInstruction = imageRefs
-      .map(({ ref }) => `Tool returned an image ref: ${ref}. To preview it for the user, write Markdown exactly as ![image](|deRef:${ref}|). For later tool calls requiring this image's base64 data URL, pass exactly "|deRef:${ref}|". Do not copy or rewrite the data URL.`)
-      .join("\n");
-  }
-  if (Array.isArray(summary.images)) {
-    summary.images = summary.images.map(image => {
-      if (!image || typeof image !== "object") return image;
-      if (typeof image.dataUrl === "string" && image.dataUrl.startsWith("data:")) {
-        const rest = { ...image };
-        delete rest.dataUrl;
-        return { ...rest, imageOmittedFromTextContext: true };
-      }
-      return image;
-    });
-  }
-
-  return summary;
-}
-
-export function collectToolResultDisplayImages(result) {
-  const images = [];
-  const pushImage = (source, options = {}) => {
-    const url = normalizeImageRefSource(source);
-    if (!url) return;
-    if (!shouldDisplayToolResultImage(url, options)) return;
-    const parsed = parseImageDataUrl(url);
-    images.push({
-      url,
-      mediaType: parsed?.mediaType,
-      kind: parsed ? "data-url" : "url"
-    });
-  };
-
-  if (Array.isArray(result?.images)) {
-    for (const image of result.images) {
-      pushImage(image?.dataUrl || image?.imageUrl || image?.url, { explicitImageField: true });
-    }
-  }
-  pushImage(result?.dataUrl, { explicitImageField: true });
-  pushImage(result?.imageUrl, { explicitImageField: true });
-  pushImage(result?.url, { explicitImageField: false });
-
-  const seen = new Set();
-  return images.filter(image => {
-    if (!image?.url || seen.has(image.url)) return false;
-    seen.add(image.url);
-    return true;
-  });
-}
-
-function shouldDisplayToolResultImage(url, { explicitImageField = false } = {}) {
-  if (!url) return false;
-  if (isBase64DataUrl(url)) return true;
-  if (/^(blob:|chrome-extension:\/\/)/i.test(url)) return explicitImageField;
-  if (/^https?:\/\//i.test(url)) {
-    return explicitImageField || isLikelyImageUrl(url);
-  }
-  return false;
-}
-
-function isLikelyImageUrl(url) {
-  try {
-    const parsed = new URL(url);
-    return /\.(png|jpe?g|gif|webp|bmp|avif|svg)(?:$|[?#])/i.test(parsed.pathname);
-  } catch {
-    return /\.(png|jpe?g|gif|webp|bmp|avif|svg)(?:$|[?#])/i.test(String(url || ""));
-  }
-}
-
-function findImageEditMcpTool(tools = []) {
-  return (tools || []).find(tool => {
-    const names = [tool?.name, tool?._toolCallName]
-      .map(name => String(name || "").trim())
-      .filter(Boolean);
-    return names.some(name => name.replace(/-/g, "_").endsWith("image_edit"));
-  }) || null;
-}
-
-function getMcpToolCallName(tool) {
-  const explicit = String(tool?._toolCallName || "").trim();
-  if (explicit) return explicit;
-  const serverName = String(tool?._serverName || "server").trim();
-  const toolName = String(tool?.name || "image_edit").trim();
-  return `mcp_${serverName}_${toolName}`;
-}
-
-function buildImageEditUserPrompt({ toolCallName, imageRef, maskRef, additionalImageRefs = [], suggestion }) {
-  const allImageRefs = [
-    imageRef,
-    ...(Array.isArray(additionalImageRefs) ? additionalImageRefs : [])
-  ].filter(Boolean);
-  const imagePlaceholders = allImageRefs.map(ref => `\`|deRef:${ref}|\``).join(", ");
-  const lines = [
-    `请使用工具 \`${toolCallName || "image_edit"}\` 修改这张图片。`,
-    `原图 ref: ${imageRef}`,
-    `调用工具时请把图片输入按这个顺序传入 images 数组: ${imagePlaceholders}。第一张是正在编辑的原图，后续图片是参考图。`
-  ];
-
-  if (maskRef) {
-    lines.push(
-      `局部修改 mask ref: ${maskRef}`,
-      `mask 参数如果需要图片数据，请传入 \`|deRef:${maskRef}|\`。`,
-      "mask 只对 images 数组第一张，也就是正在展示的原图生效；mask 是 PNG alpha mask：透明区域表示需要修改范围，不透明区域表示保持不变。请尽量只改 mask 透明区域。"
-    );
-  }
-
-  if (allImageRefs.length === 1) {
-    lines.push(`如果工具没有 images 参数，请传入 image 参数: \`|deRef:${imageRef}|\`。`);
-  } else {
-    const additionalPlaceholders = allImageRefs.slice(1).map(ref => `\`|deRef:${ref}|\``).join(", ");
-    lines.push(
-      `如果工具没有 images 参数，请传入 image 参数: \`|deRef:${imageRef}|\`，并传入 additional_images 数组: ${additionalPlaceholders}。`
-    );
-  }
-
-  lines.push(`修改建议：${String(suggestion || "").trim()}`);
-  return lines.join("\n");
-}
-
-function buildImageEditMessageRefs({
-  imageRef,
-  imageSource,
-  maskRef,
-  maskSource,
-  referenceRefs = [],
-  referenceSources = []
-}) {
-  return normalizeMessageImageRefs([
-    { ref: imageRef, dataUrl: imageSource, role: "edit_image" },
-    ...referenceRefs.map((ref, index) => ({
-      ref,
-      dataUrl: referenceSources[index],
-      role: "edit_reference"
-    })),
-    maskRef ? { ref: maskRef, dataUrl: maskSource, role: "edit_mask" } : null
-  ]);
-}
-
-function imageBlockToDataUrl(block) {
-  if (!block || typeof block !== "object") return null;
-  if (block.type === "image" && block.source?.type === "base64" && block.source.media_type && block.source.data) {
-    return `data:${block.source.media_type};base64,${block.source.data}`;
-  }
-  if (block.type === "image_url" && isBase64DataUrl(block.image_url?.url)) {
-    return block.image_url.url;
-  }
-  return null;
-}
-
-function isImageFile(file) {
-  if (!file) return false;
-  return String(file.type || "").startsWith("image/") || IMAGE_FILE_EXT.test(String(file.name || ""));
-}
-
-function getClipboardImageFiles(clipboardData) {
-  const items = Array.from(clipboardData?.items || []);
-  let imageFiles = items
-    .filter(item => item.kind === "file" && (String(item.type || "").startsWith("image/") || item.type === ""))
-    .map(item => item.getAsFile())
-    .filter(isImageFile);
-
-  const files = Array.from(clipboardData?.files || []);
-  if (imageFiles.length === 0) {
-    imageFiles = files.filter(isImageFile);
-  }
-
-  return imageFiles;
-}
-
-async function imageFileToAttachmentItem(file) {
-  const rawDataUrl = await blobToDataUrl(file);
-  const optimizedDataUrl = await optimizeImageDataUrl(rawDataUrl);
-  const parsed = parseImageDataUrl(optimizedDataUrl);
-  if (!parsed) return null;
-  return {
-    id: `att_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`,
-    type: "image",
-    dataUrl: optimizedDataUrl,
-    mediaType: parsed.mediaType,
-    fileName: file.name
-  };
-}
-
-function resolveImageRefsInValue(value, refs) {
-  if (typeof value === "string") {
-    const imageRef = normalizeImageRefToken(value);
-    if (!imageRef) return value;
-    return refs.get(imageRef) || value;
-  }
-
-  if (Array.isArray(value)) {
-    return value.map(item => resolveImageRefsInValue(item, refs));
-  }
-
-  if (value && typeof value === "object") {
-    return Object.fromEntries(
-      Object.entries(value).map(([key, child]) => [key, resolveImageRefsInValue(child, refs)])
-    );
-  }
-
-  return value;
-}
-
-function normalizeImageRefToken(value) {
-  const raw = String(value || "").trim();
-  if (!raw) return "";
-  const candidates = [raw];
-  try {
-    const decoded = decodeURIComponent(raw);
-    if (decoded !== raw) candidates.push(decoded);
-  } catch {
-    // Keep the raw candidate when it is not URL encoded.
-  }
-
-  for (const candidate of candidates) {
-    let normalized = candidate.trim();
-    while (
-      (normalized.startsWith('"') && normalized.endsWith('"')) ||
-      (normalized.startsWith("'") && normalized.endsWith("'"))
-    ) {
-      normalized = normalized.slice(1, -1).trim();
-    }
-    const derefMatch = normalized.match(IMAGE_DEREF_PATTERN);
-    if (derefMatch) return derefMatch[1];
-    if (IMAGE_REF_PATTERN.test(normalized)) return normalized;
-  }
-
-  return "";
-}
-
-function summarizeImageRefCache(sessionId, cache) {
-  const refs = cache?.refs instanceof Map ? cache.refs : new Map();
-  return {
-    sessionId,
-    count: refs.size,
-    refs: Object.fromEntries(
-      Array.from(refs.entries()).map(([ref, source]) => [ref, summarizeImageRefSource(source)])
-    )
-  };
-}
-
-function summarizeImageRefSource(source) {
-  const raw = String(source || "");
-  const parsed = parseImageDataUrl(raw);
-  return {
-    kind: parsed ? "data-url" : raw ? "url" : "empty",
-    mediaType: parsed?.mediaType || "",
-    length: raw.length,
-    preview: raw ? `${raw.slice(0, 80)}${raw.length > 80 ? "..." : ""}` : ""
-  };
-}
-
-function parseImageDataUrl(dataUrl) {
-  if (typeof dataUrl !== "string") return null;
-  const match = dataUrl.match(/^data:([^;]+);base64,(.+)$/);
-  if (!match) return null;
-  return { mediaType: match[1], data: match[2] };
-}
-
-function parseToolMessageContent(content) {
-  if (typeof content !== "string") return content;
-  try {
-    return JSON.parse(content);
-  } catch (e) {
-    return content;
-  }
-}
-
-function normalizeToolSummary(summary) {
-  if (summary && typeof summary === "object") return summary;
-  return { result: summary == null ? "" : String(summary) };
-}
-
-function buildAnthropicToolResultContentFromMessage(msg, options = {}) {
-  const summary = normalizeToolSummary(parseToolMessageContent(msg.content));
-  const parsedImages = getToolResultParsedImages(msg);
-  if (parsedImages.length === 0 || options.supportsImageInput === false) {
-    return typeof summary === "string" ? summary : JSON.stringify(summary);
-  }
-
-  return [
-    {
-      type: "text",
-      text: JSON.stringify({ ...summary, imageAttachedToToolResult: true })
-    },
-    ...parsedImages.map(parsedImage => ({
-      type: "image",
-      source: {
-        type: "base64",
-        media_type: parsedImage.mediaType,
-        data: parsedImage.data
-      }
-    }))
-  ];
-}
-
-function buildOpenAIToolResultContent(msg, options = {}) {
-  const summary = normalizeToolSummary(parseToolMessageContent(msg.content));
-  const parsedImages = getToolResultParsedImages(msg);
-  if (parsedImages.length === 0 || options.supportsImageInput === false) {
-    return typeof summary === "string" ? summary : JSON.stringify(summary);
-  }
-
-  return [
-    {
-      type: "text",
-      text:
-        `Tool result for ${msg.tool_name || "unknown tool"}: ` +
-        JSON.stringify({ ...summary, imageAttachedToToolResult: true })
-    },
-    ...parsedImages.map(parsedImage => ({
-      type: "image_url",
-      image_url: {
-        url: parsedImage.url,
-        detail: "low"
-      }
-    }))
-  ];
-}
-
-function getToolResultParsedImages(msg) {
-  const sources = Array.isArray(msg?.displayImages) && msg.displayImages.length > 0
-    ? msg.displayImages.map(image => image?.url)
-    : [msg?.displayImageUrl];
-  return sources
-    .map(url => ({ url: normalizeImageRefSource(url), parsed: parseImageDataUrl(normalizeImageRefSource(url)) }))
-    .filter(item => item.url && item.parsed)
-    .map(item => ({ url: item.url, ...item.parsed }));
-}
-
-// ==================== User Image Input Helpers ====================
-
-function buildUserMessageContent(text, images, textFiles = [], imageRefs = []) {
-  const content = [];
-
-  const textParts = [];
-  if (text && text.trim()) textParts.push(text.trim());
-  for (const item of imageRefs) {
-    if (!item?.ref) continue;
-    textParts.push(`Attached image ref: ${item.ref}. For any tool argument that requires this image's base64 data URL, pass exactly "|deRef:${item.ref}|". Do not copy or rewrite the data URL.`);
-  }
-  if (textParts.length > 0) {
-    content.push({ type: "text", text: textParts.join("\n\n") });
-  }
-
-  for (const f of textFiles) {
-    console.log(`[DEBUG] 构建消息内容 - 文件: ${f.fileName}, 文本长度: ${f.text.length}, 前100字符:`, f.text.substring(0, 100));
-    content.push({ type: "file", fileName: f.fileName, text: f.text });
-  }
-
-  for (const img of images) {
-    const parsed = parseImageDataUrl(img.dataUrl);
-    if (parsed) {
-      content.push({
-        type: "image",
-        source: { type: "base64", media_type: parsed.mediaType, data: parsed.data }
-      });
-    }
-  }
-
-  return content;
-}
-
-async function optimizeImageDataUrl(dataUrl) {
-  try {
-    const img = await new Promise((resolve, reject) => {
-      const image = new Image();
-      image.onload = () => resolve(image);
-      image.onerror = reject;
-      image.src = dataUrl;
-    });
-
-    const maxDimension = 1600;
-    const scale = Math.min(1, maxDimension / Math.max(img.width, img.height));
-    const width = Math.round(img.width * scale);
-    const height = Math.round(img.height * scale);
-
-    const canvas = document.createElement("canvas");
-    canvas.width = width;
-    canvas.height = height;
-    const ctx = canvas.getContext("2d", { alpha: false });
-    ctx.drawImage(img, 0, 0, width, height);
-
-    return canvas.toDataURL("image/jpeg", 0.7);
-  } catch (e) {
-    console.error("Image optimization failed:", e);
-    return dataUrl;
-  }
-}
-
-async function blobToDataUrl(blob) {
-  return new Promise((resolve, reject) => {
-    const reader = new FileReader();
-    reader.onload = () => resolve(reader.result);
-    reader.onerror = reject;
-    reader.readAsDataURL(blob);
-  });
-}
-
-function buildAnthropicAssistantContentFromMessage(msg, options = {}) {
-  const omitThinking = shouldOmitThinkingFromRequests(options);
-
-  if (Array.isArray(msg.content)) {
-    const contentBlocks = omitThinking ? msg.content.filter(block => !isAnthropicThinkingBlock(block)) : msg.content;
-    const hasThinkingBlocks = contentBlocks.some(isAnthropicThinkingBlock);
-    const prependedThinkingBlocks = omitThinking || hasThinkingBlocks ? [] : extractAnthropicThinkingBlocksFromMessage(msg);
-    return normalizeAnthropicAssistantContentBlocks([...prependedThinkingBlocks, ...contentBlocks]);
-  }
-
-  const blocks = omitThinking ? [] : extractAnthropicThinkingBlocksFromMessage(msg);
-  if (msg.content && typeof msg.content === "string" && msg.content.length > 0) {
-    blocks.push({ type: "text", text: msg.content });
-  }
-
-  if (Array.isArray(msg.tool_calls)) {
-    for (const tc of msg.tool_calls) {
-      const toolName = tc.function?.name || tc.name;
-      let input = tc.function?.arguments ?? tc.arguments ?? tc.args ?? {};
-      if (typeof input === "string") {
-        try { input = JSON.parse(input); } catch (e) { input = { raw: input }; }
-      }
-      if (toolName) {
-        blocks.push({
-          type: "tool_use",
-          id: tc.id || `tooluse_${toolName}_${Date.now()}`,
-          name: toolName,
-          input
-        });
-      }
-    }
-  }
-
-  return normalizeAnthropicAssistantContentBlocks(blocks);
-}
-
-function normalizeAnthropicAssistantContentBlocks(blocks) {
-  return (blocks || []).map(normalizeAnthropicAssistantContentBlock).filter(Boolean);
-}
-
-function normalizeAnthropicAssistantContentBlock(block) {
-  if (!block || typeof block !== "object") return null;
-
-  if (block.type === "text") {
-    return typeof block.text === "string" && block.text.length > 0 ? { ...block } : null;
-  }
-
-  if (block.type === "tool_use") {
-    if (!block.name) return null;
-    return {
-      ...block,
-      input: normalizeAnthropicToolUseInput(block.input)
-    };
-  }
-
-  if (block.type === "thinking") {
-    const signature = typeof block.signature === "string" ? block.signature : "";
-    if (!signature) return null;
-    return {
-      ...block,
-      thinking: typeof block.thinking === "string" ? block.thinking : ""
-    };
-  }
-
-  if (block.type === "redacted_thinking") {
-    return block.data ? { ...block } : null;
-  }
-
-  return { ...block };
-}
-
-function normalizeAnthropicToolUseInput(input) {
-  if (input && typeof input === "object" && !Array.isArray(input)) return input;
-  if (typeof input === "string") {
-    try { return JSON.parse(input); } catch (error) { return { raw: input }; }
-  }
-  return input ?? {};
-}
-
-function extractAnthropicThinkingBlocksFromMessage(msg) {
-  const blocks = [];
-
-  if (Array.isArray(msg?.thinking_blocks)) {
-    blocks.push(...msg.thinking_blocks);
-  }
-
-  const providerReasoningBlocks = msg?.provider_specific_fields?.reasoningContentBlocks;
-  if (Array.isArray(providerReasoningBlocks)) {
-    for (const block of providerReasoningBlocks) {
-      const reasoningText = block?.reasoningText;
-      if (reasoningText?.signature) {
-        blocks.push({
-          type: "thinking",
-          thinking: reasoningText.text || reasoningText.thinking || "",
-          signature: reasoningText.signature
-        });
-        continue;
-      }
-
-      const redacted = block?.redactedContent || block?.redactedThinking || block?.redacted_thinking;
-      if (redacted?.data) {
-        blocks.push({ type: "redacted_thinking", data: redacted.data });
-      }
-    }
-  }
-
-  return normalizeAnthropicAssistantContentBlocks(blocks).filter(isAnthropicThinkingBlock);
-}
-
-function isAnthropicThinkingBlock(block) {
-  return block?.type === "thinking" || block?.type === "redacted_thinking";
-}
-
-function buildOpenAIAssistantMessageFromAnthropic(msg, options = {}) {
-  if (!Array.isArray(msg.content)) return buildOpenAIAssistantMessageForApi(msg, options);
-
-  const omitThinking = shouldOmitThinkingFromRequests(options);
-  const textParts = [];
-  const reasoningParts = [];
-  const toolCalls = [];
-
-  for (const block of msg.content) {
-    if (!block) continue;
-    if (block.type === "text" && block.text) {
-      textParts.push(block.text);
-    } else if (!omitThinking && block.type === "thinking" && block.thinking) {
-      reasoningParts.push(block.thinking);
-    } else if (block.type === "tool_use" && block.name) {
-      toolCalls.push({
-        id: block.id || `toolcall_${block.name}_${Date.now()}`,
-        type: "function",
-        function: {
-          name: block.name,
-          arguments: JSON.stringify(block.input || {})
-        }
-      });
-    }
-  }
-
-  const apiMessage = {
-    role: "assistant",
-    content: textParts.length > 0 ? textParts.join("") : null,
-    ...(toolCalls.length > 0 ? { tool_calls: toolCalls } : {})
-  };
-  if (reasoningParts.length > 0) {
-    apiMessage.reasoning_content = reasoningParts.join("");
-  }
-  return copyOpenAIProviderMetadataForApi(msg, copyOpenAIReasoningFieldsForApi(msg, apiMessage, options), options);
-}
-
-function buildOpenAIAssistantMessageForApi(msg, options = {}) {
-  if (Array.isArray(msg.content)) {
-    return buildOpenAIAssistantMessageFromAnthropic(msg, options);
-  }
-
-  const apiMessage = {
-    role: "assistant",
-    content: msg.content ?? null
-  };
-  if (Array.isArray(msg.tool_calls) && msg.tool_calls.length > 0) {
-    apiMessage.tool_calls = msg.tool_calls;
-  }
-  return copyOpenAIProviderMetadataForApi(msg, copyOpenAIReasoningFieldsForApi(msg, apiMessage, options), options);
-}
-
-function copyOpenAIReasoningFieldsForApi(source, target, options = {}) {
-  if (shouldOmitThinkingFromRequests(options)) return target;
-
-  const reasoningContent = getOpenAIReasoningContentForApi(source);
-  if (reasoningContent != null) {
-    target.reasoning_content = reasoningContent;
-  }
-
-  for (const field of ["reasoning", "reasoning_details", "thinking"]) {
-    const value = source?.[field];
-    if (value == null) continue;
-    if (typeof value === "string" && value.length === 0) continue;
-    if (Array.isArray(value) && value.length === 0) continue;
-    target[field] = value;
-  }
-  return target;
-}
-
-function getOpenAIReasoningContentForApi(msg) {
-  for (const field of ["reasoning_content", "reasoning", "thinking"]) {
-    const value = msg?.[field];
-    if (typeof value === "string" && value.length > 0) return value;
-  }
-  return null;
-}
-
-function copyOpenAIProviderMetadataForApi(source, target, options = {}) {
-  if (shouldOmitThinkingFromRequests(options)) return target;
-  if (normalizeApiType(options.apiType) !== API_TYPES.OPENAI_RESPONSES) return target;
-  if (Array.isArray(source?._responsesReasoningItems) && source._responsesReasoningItems.length > 0) {
-    target._responsesReasoningItems = source._responsesReasoningItems;
-  }
-  return target;
-}
-
-function buildOpenAIApiMessages(messages, options = {}) {
-  const apiMessages = [];
-
-  for (let i = 0; i < messages.length; i++) {
-    const msg = messages[i];
-
-    if (msg.role === "error") continue;
-
-    if (msg.role === "assistant" && Array.isArray(msg.tool_calls) && msg.tool_calls.length > 0) {
-      const followingToolMessages = [];
-      let j = i + 1;
-      while (j < messages.length && messages[j]?.role === "tool") {
-        followingToolMessages.push(messages[j]);
-        j += 1;
-      }
-
-      apiMessages.push(buildOpenAIAssistantMessageForApi(msg, options));
-      apiMessages.push(...followingToolMessages.map(toolMsg => ({
-        role: "tool",
-        tool_call_id: toolMsg.tool_call_id,
-        content: buildOpenAIToolResultContent(toolMsg, options)
-      })));
-
-      i = j - 1;
-      continue;
-    }
-
-    if (msg.role === "assistant") {
-      apiMessages.push(buildOpenAIAssistantMessageForApi(msg, options));
-      continue;
-    }
-
-    if (msg.role === "user") {
-      if (Array.isArray(msg.content)) {
-        const openaiContent = [];
-        for (const block of msg.content) {
-          if (block.type === "text") {
-            openaiContent.push({ type: "text", text: block.text });
-          } else if (block.type === "file") {
-            openaiContent.push({ type: "text", text: `[Attached file: ${block.fileName}]\n${block.text}` });
-          } else if (block.type === "image" && block.source && options.supportsImageInput !== false) {
-            const dataUrl = `data:${block.source.media_type};base64,${block.source.data}`;
-            openaiContent.push({ type: "image_url", image_url: { url: dataUrl, detail: "low" } });
-          }
-        }
-        apiMessages.push({ role: "user", content: openaiContent });
-        continue;
-      }
-    }
-
-    if (msg.role === "tool") {
-      apiMessages.push({
-        role: "tool",
-        tool_call_id: msg.tool_call_id,
-        content: buildOpenAIToolResultContent(msg, options)
-      });
-      continue;
-    }
-
-    apiMessages.push(buildPlainApiMessage(msg, options));
-  }
-
-  return apiMessages;
-}
-
-function buildAnthropicApiMessages(messages, options = {}) {
-  const apiMessages = [];
-
-  for (let i = 0; i < messages.length; i++) {
-    const msg = messages[i];
-
-    if (msg.role === "error") continue;
-
-    if (msg.role === "tool") {
-      const blocks = [];
-      while (i < messages.length && messages[i]?.role === "tool") {
-        const toolMsg = messages[i];
-        blocks.push({
-          type: "tool_result",
-          tool_use_id: toolMsg.tool_call_id,
-          content: buildAnthropicToolResultContentFromMessage(toolMsg, options)
-        });
-        i += 1;
-      }
-      i -= 1;
-      if (blocks.length > 0) {
-        apiMessages.push({ role: "user", content: blocks });
-      }
-      continue;
-    }
-
-    if (msg.role === "assistant") {
-      const content = buildAnthropicAssistantContentFromMessage(msg, options);
-      if (content.length === 0) continue;
-      apiMessages.push({ role: "assistant", content });
-      continue;
-    }
-
-    if (msg.role === "user" && Array.isArray(msg.content)) {
-      const anthropicContent = msg.content
-        .filter(block => !(block.type === "image" && options.supportsImageInput === false))
-        .map(block => {
-          if (block.type === "file") {
-            const result = { type: "text", text: `[Attached file: ${block.fileName}]\n${block.text}` };
-            console.log(`[DEBUG] Anthropic API - 文件转换: ${block.fileName}, 原始长度: ${block.text.length}, 转换后长度: ${result.text.length}`);
-            return result;
-          }
-          return block;
-        });
-      apiMessages.push({ role: "user", content: anthropicContent });
-      continue;
-    }
-
-    apiMessages.push(buildPlainApiMessage(msg, options));
-  }
-
-  return apiMessages;
-}
-
-function buildPlainApiMessage(msg, options = {}) {
-  if (!msg || typeof msg !== "object") return msg;
-
-  const apiMessage = { ...msg };
-  for (const field of [
-    "sentAt",
-    "durationMs",
-    "displayImageUrl",
-    "displayImages",
-    "displayImageMediaType",
-    "displayImageOmitFromRequests",
-    "_usageApiType",
-    "_usageModel",
-    "displayContent",
-    "injectedUserContext",
-    "imageRefs"
-  ]) {
-    delete apiMessage[field];
-  }
-  if (shouldOmitThinkingFromRequests(options)) {
-    for (const field of [
-      "thinking_blocks",
-      "provider_specific_fields",
-      "reasoning_content",
-      "reasoning",
-      "reasoning_details",
-      "thinking",
-      "_responsesReasoningItems"
-    ]) {
-      delete apiMessage[field];
-    }
-  }
-  return apiMessage;
-}
-
-function shouldOmitThinkingFromRequests(options = {}) {
-  return options.omitThinkingFromRequests === true;
-}
-
-function buildApiMessages(apiType, messages, options = {}) {
-  const normalizedApiType = normalizeApiType(apiType);
-  if (normalizedApiType === API_TYPES.ANTHROPIC) {
-    return buildAnthropicApiMessages(messages, options);
-  }
-  return buildOpenAIApiMessages(messages, { ...options, apiType: normalizedApiType });
-}
-
-function buildPlatformSystemPrompt(platformInfo) {
-  if (!platformInfo?.os) {
-    return "";
-  }
-
-  const parts = [`Current operating system: ${platformInfo.os}`];
-  if (platformInfo.arch) parts.push(`architecture: ${platformInfo.arch}`);
-  if (platformInfo.nacl_arch) parts.push(`nacl_arch: ${platformInfo.nacl_arch}`);
-
-  return `Environment:\n- ${parts.join("; ")}.\n\n`;
-}
-
-async function loadSkillsIndexFromSkillStation(serverUrl) {
-  const normalizedServerUrl = normalizeSkillStationUrl(serverUrl);
-  const connection = await connectMcpServer(normalizedServerUrl, {});
-  if (connection.error) {
-    throw new Error(connection.error);
-  }
-  const resources = await listMcpResources(normalizedServerUrl);
-  const skillsIndex = resources.find(resource => resource?.uri === "skills://index");
-  if (!skillsIndex) {
-    throw new Error("skill-bridge 未暴露 skills://index 资源");
-  }
-
-  const resourceResult = await readMcpResource(normalizedServerUrl, {}, "skills://index");
-  return parseLoadedSkillsResponse(extractResourceText(resourceResult));
-}
-
-function parseLoadedSkillsResponse(text) {
-  const payloadText = extractJsonPayload(text);
-  let payload;
-
-  try {
-    payload = JSON.parse(payloadText);
-  } catch (error) {
-    throw new Error("Skills 索引返回的不是合法 JSON");
-  }
-
-  if (payload?.error) {
-    throw new Error(String(payload.error));
-  }
-
-  const rawSkills = Array.isArray(payload)
-    ? payload
-    : (Array.isArray(payload?.skills) ? payload.skills : null);
-
-  if (!rawSkills) {
-    throw new Error("Skills 索引缺少 skills 数组");
-  }
-
-  return rawSkills
-    .map(skill => ({
-      path: String(skill?.directoryName || skill?.path || "").trim().replace(/\\/g, "/").replace(/^\.?\//, ""),
-      name: String(skill?.name || "").trim(),
-      description: String(skill?.description || "").trim(),
-      header: {
-        ...(skill?.metadata && typeof skill.metadata === "object" ? skill.metadata : {}),
-        ...(skill?.header && typeof skill.header === "object" ? skill.header : {})
-      }
-    }))
-    .filter(skill => !!skill.path);
-}
-
-async function loadSkillStationTools(serverUrl, bridgeToolSettings = {}) {
-  const normalizedServerUrl = normalizeSkillStationUrl(serverUrl);
-  const result = await connectMcpServer(normalizedServerUrl, {});
-  if (result.error) {
-    throw new Error(result.error);
-  }
-
-  const tools = Array.isArray(result.tools) ? result.tools : [];
-  const hasGetSkillDetail = tools.some(tool => tool?.name === "get_skill_detail");
-  if (!hasGetSkillDetail) {
-    throw new Error("skill-bridge 缺少 get_skill_detail 工具");
-  }
-
-  return tools.map(tool => ({
-    ...tool,
-    _serverId: "skill_bridge",
-    _serverName: "skill_bridge",
-    _serverUrl: normalizedServerUrl,
-    _serverHeaders: {},
-    _dangerous: resolveSkillBridgeToolDangerous(tool.name, bridgeToolSettings),
-    _toolCallName: `mcp_skill_bridge_${tool.name}`
-  }));
-}
-
-function resolveSkillBridgeToolDangerous(toolName, bridgeToolSettings = {}) {
-  const normalizedToolName = String(toolName || "").trim();
-  const explicitDangerous = bridgeToolSettings?.[normalizedToolName]?.dangerous;
-  if (explicitDangerous != null) {
-    return !!explicitDangerous;
-  }
-  return normalizedToolName === "shell";
-}
-
-function normalizeSkillStationUrl(serverUrl) {
-  return String(serverUrl || "").trim();
-}
-
-function extractResourceText(resourceResult) {
-  const contents = Array.isArray(resourceResult?.contents) ? resourceResult.contents : [];
-  const texts = contents
-    .map(item => item?.text)
-    .filter(text => typeof text === "string" && text.trim().length > 0);
-
-  if (texts.length === 0) {
-    throw new Error("skills://index 返回为空");
-  }
-
-  return texts.join("\n");
-}
-
-function mergeMcpToolLists(primaryTools, secondaryTools) {
-  const map = new Map();
-  for (const tool of [...(primaryTools || []), ...(secondaryTools || [])]) {
-    if (!tool?._toolCallName) continue;
-    map.set(tool._toolCallName, tool);
-  }
-  return [...map.values()];
-}
-
-function extractJsonPayload(text) {
-  const raw = String(text || "").trim();
-  if (!raw) {
-    throw new Error("Skills 索引返回为空");
-  }
-
-  const fenced = raw.match(/```(?:json)?\s*([\s\S]*?)```/i);
-  if (fenced?.[1]) return fenced[1].trim();
-
-  if (raw.startsWith("{") || raw.startsWith("[")) return raw;
-
-  const firstArrayStart = raw.indexOf("[");
-  const lastArrayEnd = raw.lastIndexOf("]");
-  if (firstArrayStart >= 0 && lastArrayEnd > firstArrayStart) {
-    return raw.slice(firstArrayStart, lastArrayEnd + 1);
-  }
-
-  const firstObjectStart = raw.indexOf("{");
-  const lastObjectEnd = raw.lastIndexOf("}");
-  if (firstObjectStart >= 0 && lastObjectEnd > firstObjectStart) {
-    return raw.slice(firstObjectStart, lastObjectEnd + 1);
-  }
-
-  throw new Error("未找到可解析的 JSON 输出");
-}
